refactor: clean up webpack and legacy code
