+++ .babelrc
@@ -10,9 +10,6 @@
     "stage-2"
   ],
   "env": {
-    "development": {
-      "plugins": "react-hot-loader/babel"
-    },
     "test": {
       "plugins": ["transform-es2015-modules-commonjs"]
     }
