+++ .eslintignore
@@ -8,7 +8,3 @@ src/client/vendor
 assets/
 public/
 busy.server.js
-
-# TODO: Eventually should be linted as well
-# Webpack
-webpack/
