+++ package.json
@@ -17,7 +17,7 @@
     "prettier-check-all": "node ./scripts/prettier/tasks/checkFiles.js all",
     "prettier-file": "node ./scripts/prettier/tasks/formatFile.js",
     "dev-server": "run-p dev-server:auth-server dev-server:client",
-    "dev-server:client": "webpack-dev-server --config ./webpack/webpack-dev-server.js",
+    "dev-server:client": "webpack-dev-server --config ./webpack/webpack.dev.config.js",
     "dev-server:auth-server": "node ./src/server/auth-server.js",
     "start": "cross-env NODE_ENV=production npm run run:server",
     "test": "jest",
@@ -27,7 +27,7 @@
     "run:server": "nodemon --watch busy.server.js busy.server.js",
     "build:server:dev": "cross-env NODE_ENV=development webpack --config ./webpack/webpack.server.config.js",
     "build:server": "cross-env NODE_ENV=production webpack -p --config ./webpack/webpack.server.config.js",
-    "build:client": "webpack -p --config ./webpack/webpack.prod.config.js",
+    "build:client": "cross-env NODE_ENV=production webpack -p --config ./webpack/webpack.config.js",
     "build": "cross-env NODE_ENV=production npm run build:client && npm run build:server",
     "heroku-prebuild": "npm i -g pm2 && pm2 install pm2-logrotate",
     "heroku-postbuild": "npm run build"
@@ -90,7 +90,6 @@
     "matchmedia": "^0.1.2",
     "nodemon": "^1.9.2",
     "npm-run-all": "^4.0.2",
-    "postcss-flexbugs-fixes": "^3.0.0",
     "postcss-loader": "^2.0.8",
     "prop-types": "^15.5.10",
     "react": "^16.2.0",
@@ -101,7 +100,6 @@
     "react-dropzone": "^4.1.3",
     "react-helmet": "^5.0.2",
     "react-highcharts": "^15.0.0",
-    "react-hot-loader": "^3.1.1",
     "react-hotkeys": "^0.10.0",
     "react-image-lightbox": "^4.0.0",
     "react-infinite-scroller": "^1.0.15",
@@ -157,8 +155,7 @@
     "minimatch": "^3.0.4",
     "prettier": "^1.9.2",
     "react-test-renderer": "^15.6.1",
-    "webpack-dev-server": "~2.9.3",
-    "webpack-hot-middleware": "^2.20.0"
+    "webpack-dev-server": "~2.9.3"
   },
   "jest": {
     "moduleNameMapper": {
