+++ src/client/post/Post.js
@@ -114,8 +114,8 @@ export default class Post extends React.Component {
   }
 
   componentWillUnmount() {
-    if (process.env.IS_BROWSER) {
-      global.document.title = 'Busy';
+    if (typeof window !== 'undefined') {
+      window.document.title = 'Busy';
     }
   }
