+++ src/client/store.js
@@ -4,21 +4,16 @@ import { applyMiddleware, createStore, compose } from 'redux';
 import { routerMiddleware } from 'react-router-redux';
 import steemAPI from './steemAPI';
 import { history } from './routes';
-import { mountResponsive } from './vendor/responsive';
 import errorMiddleware from './errorMiddleware';
 import reducers from './reducers';
 
 export default steemConnectAPI => {
   let preloadedState;
-  if (process.env.IS_BROWSER) {
+  if (typeof window !== 'undefined') {
     /* eslint-disable no-underscore-dangle */
     preloadedState = window.__PRELOADED_STATE__;
     delete window.__PRELOADED_STATE__;
     /* eslint-enable no-underscore-dangle */
-  }
-
-  if (process.env.IS_BROWSER && process.env.NODE_ENV !== 'production') {
-    window.steemAPI = steemAPI;
   }
 
   const middleware = [
@@ -34,7 +29,7 @@ export default steemConnectAPI => {
   ];
 
   let enhancer;
-  if (process.env.IS_BROWSER) {
+  if (typeof window !== 'undefined') {
     // eslint-disable-next-line no-underscore-dangle
     const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
     enhancer = composeEnhancers(applyMiddleware(...middleware));
@@ -43,6 +38,5 @@ export default steemConnectAPI => {
   }
 
   const store = createStore(reducers, preloadedState, enhancer);
-  mountResponsive(store);
   return store;
 };
