--- webpack/index.js
@@ -1,24 +1,0 @@
-'use strict';
-
-const webpack = require('webpack');
-const webpackMiddleware = require('webpack-dev-middleware');
-const webpackHotMiddleware = require('webpack-hot-middleware');
-const webpackConfig = require('./webpack.config.js');
-
-module.exports = function (app) {
-  const compiler = webpack(webpackConfig);
-  const middleware = webpackMiddleware(compiler, {
-    publicPath: webpackConfig.output.publicPath,
-    contentBase: 'src',
-    stats: {
-      colors: true,
-      hash: false,
-      timings: true,
-      chunks: false,
-      chunkModules: false,
-      modules: false
-    }
-  });
-  app.use(middleware);
-  app.use(webpackHotMiddleware(compiler));
-};
