--- webpack/webpack.prod.config.js
@@ -1,3 +1,0 @@
-module.exports = require('./makeConfig')({
-  isDevelopment: false,
-});
